requirements.txt
Pillow==11.1.0
playsound==1.3.0
